+++ package_monitor/tests/test_tasks.py
@@ -7,7 +7,7 @@ from .. import tasks
 
 class TestTasks(TestCase):
     @patch("package_monitor.tasks.Distribution.objects.update_all")
-    def test_shoud_update_all_distributions(self, update_all):
+    def test_should_update_all_distributions(self, update_all):
         # when
         tasks.update_distributions()
         # then