+++ CHANGELOG.md
@@ -13,6 +13,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+## [1.6.0] - 2022-08-30
+
+### Added
+
+- Ability to exclude distribution packages via new setting
+- Ability to exclude editable packages via new setting
+
+### Changed
+
+- No longer shows editable packages by default
+
 ## [1.5.0] - 2022-08-02
 
 ### Update notes